requirements.txt
annotated-types==0.7.0
aiohttp==3.9.1
anyio==3.7.1
APScheduler==3.10.4
certifi==2024.2.2
click==8.1.7
exceptiongroup==1.2.2
fastapi>=0.100.0
gunicorn==21.2.0  # Added for production server
httpx>=0.23.0
h11==0.14.0
idna==2.10
Jinja2==3.1.4
MarkupSafe==3.0.2
openai>=1.3.7
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-settings==0.2.2
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
starlette>=0.27.0
typing_extensions==4.12.2
uvicorn[standard]==0.32.0  # Updated to include standard extras
# pinecone-client>=3.0.0
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
supabase>=2.0.0